Remove superfluous DegreeOfMatrixGroup method
DegreeOfMatrixGroup is a synonym for DimensionOfMatrixGroup -- that means they
are actually identical. Installing an operation as a method for itself will cause a segfault
if that method ever is invoked Luckily, for attribute storing reps (which currently is used
by all our matrix groups), that method never gets invoked anyway precisely because the
two attributes are identical: if HasDimensionOfMatrixGroup(G) gives true, then the system
getter (which has a very high rank) is used to retrieve the value.

Still, it seems best to remove this.